Undo renaming
